Use long integer for set_buffer_address_4_patch

+++ MHD/C_src/GLSL/vartex_array_object_gl.h
@@ -23,7 +23,7 @@ struct VAO_ids{
 
     unsigned int id_index;
 	
-    unsigned int npoint_draw;
+    unsigned long npoint_draw;
 };
 
 /* Prototypes */
