+++ MHD/C_src/KEMO_GL/set_PSF_isolines_to_buf.c
@@ -45,12 +45,12 @@ int find_start_positive_lines(int n_isoline, struct colormap_params *cmap_s){
     return ist_positive_line;
 }
 
-static int add_num_PSF_isolines(int ist_patch, int ist, int ied, 
-								struct psf_data *psf_s, struct psf_menu_val *psf_m){
+static long add_num_PSF_isolines(long ist_patch, int ist, int ied,
+                                 struct psf_data *psf_s, struct psf_menu_val *psf_m){
 	int j;
 	double v_line;
 	
-	int num_patch = ist_patch;
+	long num_patch = ist_patch;
 	for (j = ist; j < ied; j++){
 		v_line = cal_isoline_value(j, psf_m->n_isoline,
 								   psf_m->cmap_psf_comp[psf_m->icomp_draw_psf]);
@@ -91,8 +91,8 @@ static long set_PSF_isolines_to_buf(const long ist_patch, int ist, int ied,
 }
 
 
-int count_PSF_all_isolines_to_buf(struct psf_data *psf_s, struct psf_menu_val *psf_m){
-	int num_patch = 0;
+long count_PSF_all_isolines_to_buf(struct psf_data *psf_s, struct psf_menu_val *psf_m){
+	long num_patch = 0;
 	if(psf_m->draw_psf_grid  != 0){
 		psf_m->ist_positive_line = find_start_positive_lines(psf_m->n_isoline,
 								psf_m->cmap_psf_comp[psf_m->icomp_draw_psf]);
