+++ MHD/C_src/KEMO_GL/set_PSF_patches_to_buf.h
@@ -26,7 +26,7 @@ void set_psf_textures_to_buf(int ist_psf, int ied_psf, struct psf_data **psf_s,
 void set_psf_map_to_buf(int ist_psf, int ied_psf, struct psf_data **psf_s, 
 			struct kemo_array_control *psf_a, struct gl_strided_buffer *strided_buf);
 
-int count_psf_arrows_to_buf(int ncorner, struct psf_data *psf_s, struct psf_menu_val *psf_m);
+long count_psf_arrows_to_buf(int ncorner, struct psf_data *psf_s, struct psf_menu_val *psf_m);
 long set_psf_arrows_to_buf(long ist_patch, int ncorner, 
                            struct psf_data *psf_s, struct psf_menu_val *psf_m,
                            struct gl_strided_buffer *strided_buf);
